migrate implementation to new component: notes/NotesEditor.js
